+++ MANIFEST.in
@@ -103,6 +103,8 @@ include psutil/arch/openbsd/proc.c
 include psutil/arch/openbsd/proc.h
 include psutil/arch/openbsd/socks.c
 include psutil/arch/openbsd/socks.h
+include psutil/arch/openbsd/users.c
+include psutil/arch/openbsd/users.h
 include psutil/arch/osx/cpu.c
 include psutil/arch/osx/cpu.h
 include psutil/arch/osx/disk.c
