+++ psutil/arch/openbsd/users.h
@@ -6,6 +6,4 @@
 
 #include <Python.h>
 
-#if defined(PSUTIL_LINUX) || defined(PSUTIL_FREEBSD) || defined(PSUTIL_NETBSD)
 PyObject *psutil_users(PyObject* self, PyObject* args);
-#endif
